"프로토 타입 세팅 + login"
