+++ app/src/main/java/com/example/umc/Diet/DietActivity.kt
@@ -1,4 +1,4 @@
-package com.example.umcproject
+package com.example.umc.Diet
 
 import android.content.Intent
 import android.graphics.Color
@@ -11,32 +11,35 @@ import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import com.example.umc.MainActivity
 import com.example.umc.R
+import com.example.umc.Onboarding.PriceActivity
 
-class SubscribeActivity : AppCompatActivity() {
+class DietActivity : AppCompatActivity() {
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_subscribe)
+        setContentView(R.layout.activity_diet)
 
         val dotsLayout = findViewById<LinearLayout>(R.id.dotsLayout)
-        updateDots(dotsLayout, 2) // 세 번째 점 활성화
+        updateDots(dotsLayout, 0) // 첫 번째 점 활성화
 
-        // "이거먹자 구독" 텍스트뷰 색상 변경
+        // "이거먹자 식단" 텍스트뷰 색상 변경
         val textView3 = findViewById<TextView>(R.id.textView3)
         setDietTextColor(textView3)
 
-        // 다음 버튼 클릭 이벤트 설정
+        //다음 버튼
         val nextButton = findViewById<Button>(R.id.nextButton)
         nextButton.setOnClickListener {
-            // MyActivity로 이동
-            val intent = Intent(this, MyActivity::class.java)
+            val intent = Intent(this, PriceActivity::class.java)
             startActivity(intent)
         }
 
         //뒤로가기 버튼
         val backButton = findViewById<ImageButton>(R.id.backButton)
         backButton.setOnClickListener {
-            val intent = Intent(this, PriceActivity::class.java)
+            val intent = Intent(this, MainActivity::class.java)
             startActivity(intent)
         }
     }
@@ -53,7 +56,7 @@ class SubscribeActivity : AppCompatActivity() {
     }
 
     private fun setDietTextColor(textView: TextView) {
-        val text = "이거먹자 구독"
+        val text = "이거먹자 식단"
         val spannableString = SpannableString(text)
 
         // "이거먹자"에 색상 적용 (#5C5C5C)
@@ -63,7 +66,7 @@ class SubscribeActivity : AppCompatActivity() {
             Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
         )
 
-        // "구독"에 색상 적용 (#FF7300)
+        // "식단"에 색상 적용 (#FF7300)
         spannableString.setSpan(
             ForegroundColorSpan(Color.parseColor("#FF7300")),
             5, 7, // "식단"의 인덱스 범위
