+++ app/src/main/java/com/example/umc/Diet/HomeContainerFragment.kt
@@ -1,10 +1,11 @@
-package com.example.umc
+package com.example.umc.Diet
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.umc.R
 import com.example.umc.databinding.FragmentHomeContainerBinding
 import com.google.android.material.tabs.TabLayoutMediator
