+++ app/src/main/java/com/example/umc/Main/MainActivity.kt
@@ -1,12 +1,11 @@
 package com.example.umc
 
-import android.content.Intent
 import android.os.Bundle
-import android.widget.Button
 import android.view.View
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
+import com.example.umc.Diet.DietDetailFragment
+import com.example.umc.Diet.HomeContainerFragment
 import com.example.umc.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
